% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_organ_function.R
\name{assess_organ}
\alias{assess_organ}
\title{Assess Organ Failure Function}
\usage{
assess_organ(
  .data,
  method = c("SIRS", "SOFA", "qSOFA"),
  patientid,
  time,
  period = 1,
  vitals
)
}
\arguments{
\item{.data}{The dataset you are working with.}

\item{method}{A character value indicating what method you would like to use to score 
organ failure.}

\item{patientid}{A character or numeric value unique to each patient.}

\item{time}{A POSIXct value that represents when a vital was taken.}

\item{period}{A numeric value that indicates how long one vital is good for, default is 
one hour. Note: This will only fill in missing vitals.}

\item{vitals}{A character vector denoting the certain vitals necessary for each organ 
failure score.}
}
\value{
Returns a data frame or tibble with a new column that denotes the desired score
}
\description{
Finds the current organ failure score.
}
