% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cardiac_sofa_score.R
\name{calc_card_sofa}
\alias{calc_card_sofa}
\title{SOFA Cardiac calculation function}
\usage{
calc_card_sofa(.data, SBP, DBP, Vasopressor, Vasopressor_dose)
}
\arguments{
\item{.data}{Your dataset}

\item{SBP}{A numeric column that represents the systolic blood pressure}

\item{DBP}{A numeric column that represents the dyastolic blood pressure}

\item{Vasopressor}{A character column that represents a list of vasopressors 
given to a patient}

\item{Vasopressor_dose}{A numeric column that represents the dosage of vasopressor 
given, needs to be in mcg/kg/min}
}
\description{
Finds the current cardiac score for the SOFA score
}
\examples{
\dontrun{
cardiovascular_flag = calc_card_sofa(.data, sbp, dbp, vasopressor, vasopressor_dose)}
}
