% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_time_between.R
\name{calc_time_between}
\alias{calc_time_between}
\title{Time Between Function}
\usage{
calc_time_between(
  dataset,
  begining_time,
  end_time,
  new_name = "time_diff",
  unitx = "days"
)
}
\arguments{
\item{dataset}{The dataset you are working with.}

\item{begining_time}{POSIXct variable that denotes the start of the time interval.}

\item{end_time}{POSIXct variable that that denotes the end of the time interval.}

\item{new_name}{Character variable that indicates what you want to the time 
difference variable to be called.}

\item{unitx}{Character variable that indicates what units to return time difference, 
default is days.}
}
\value{
Gives a dataset with a new variable consisting of the difference between 
two time periods
}
\description{
Finds the difference between two timestamps.
}
\examples{
\dontrun{
data <- data \%>\% calc_time_between(culture_start, culture_end, unitx = "days")
data <- calc_time_between(data, lactate_start, lactate_end, unitx = "hours")
}
}
