% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_culture_window.R
\name{find_bx_window}
\alias{find_bx_window}
\title{Function that finds the blood culture window according to the CDC definition of Sepsis.}
\usage{
find_bx_window(.data, timestamp_variable, blood_culture_time)
}
\arguments{
\item{.data}{The dataset you are working with.}

\item{timestamp_variable}{The time for the variable you want to test it it occurs within the
blood culture window}

\item{blood_culture_time}{The timestamp of your blood culture value}
}
\value{
Gives a dataset with a new variable consisting of the difference between two time periods
}
\description{
According to the CDC, in order for sepsis to take place, organ dysfunction and antibiotic 
administration must have taken place within 2 days before or 2 days after blood cultures 
being drawn.
}
\examples{
\dontrun{
result <- find_bx_window(data, service_time, culture_draw_time)
}
}
