% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_hospital_day_function.R
\name{find_hospital_day}
\alias{find_hospital_day}
\alias{find_hospital_day.default}
\title{Function to calculate what hospital day event occurred}
\usage{
find_hospital_day(.data, event, patientid, admission_day)

\method{find_hospital_day}{default}(.data, event, patientid, admission_day)
}
\arguments{
\item{.data}{Your data}

\item{event}{Date object to be used against admit time}

\item{patientid}{A column denoting a unique patient identifier}

\item{admission_day}{First day that a patient arrived at the hospital}
}
\value{
Returns a dataset with a new column denoting the hospital day
for each date given
}
\description{
Function to calculate what hospital day event occurred

default method
}
\examples{
\dontrun{
result <- find_hospital_day(blood_abx, Service_Day)
}
}
