% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organ_dysfunction_scores.R
\name{find_qsofa}
\alias{find_qsofa}
\title{qSOFA calculation function}
\usage{
find_qsofa(
  .data,
  patientid,
  time,
  period = 1,
  vitals = c(RR = NA, SBP = NA, GCS = NA)
)
}
\arguments{
\item{.data}{Your dataset}

\item{patientid}{A value indicating the unique patient id, usually an encounter number.}

\item{time}{A POSIXct value indicating the timestamp for when vitals where taken.}

\item{period}{A numeric value that indicates how long one vital is good for,
default is one hour. Note: This will only fill in missing vitals.}

\item{vitals}{A character vector denoting what columns represent the vitals 
needed to calculate qSOFA}
}
\description{
Finds the current qSOFA score
}
\examples{
\dontrun{
result <- find_qsofa(data, patientid = Encounter, time = Service_Time, 
vitals = c("RR" = "rr", "SBP" = "sbp", "GCS" = "gcs"))
}
}
