% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organ_dysfunction_scores.R
\name{find_sirs}
\alias{find_sirs}
\title{SIRS calculation function}
\usage{
find_sirs(
  .data,
  patientid,
  time,
  period = 1,
  vitals = c(Temperature = NA, RR = NA, HR = NA, WBC = NA)
)
}
\arguments{
\item{.data}{The dataset you are working with.}

\item{patientid}{A value indicating the unique patient id, usually an encounter number.}

\item{time}{A POSIXct value indicating the timestamp for when vitals where taken.}

\item{period}{A numeric value that indicates how long one vital is good for, default is 
one hour. Note: This will only fill in missing vitals.}

\item{vitals}{A character vector denoting what columns represent the vitals needed 
to calculate SIRS, it is important to have temperature in Celsius and WBC in mmm^3}
}
\description{
Finds the current organ failure score using SIRS (Systematic Inflammatory Response Syndrom)
}
\examples{
\dontrun{
dataset <- find_sirs(dataset, Encounter, `Vital Timestamp`, 
 vitals = c("rr" = "RR", "temperature" = "Temperature", "hr" = "HR", "wbc" = "Wbc"))
}
}
