% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organ_dysfunction_scores.R
\name{find_sofa}
\alias{find_sofa}
\title{SOFA calculation function}
\usage{
find_sofa(
  .data,
  patientid,
  time,
  period = 1,
  vitals = c(PaO2 = NA, FiO2 = NA, Platelets = NA, Bilirubin = NA, GCS = NA, Creatinine
    = NA, SBP = NA, DBP = NA, Vasopressor = NA, Vasopressor_Dose = NA)
)
}
\arguments{
\item{.data}{The dataset you are working with. Must contain 
PaO2, FiO2, bilirubin, vasopressor, vasopressoage dosage, creatinine, glasgow coma scale (gcs), 
systolic blood pressure (sbc), diastolic blood pressure (dbc), and platelet count columns.}

\item{patientid}{A value indicating the unique patient id, usually an encounter number.}

\item{time}{A POSIXct value indicating the timestamp for when vitals where taken.}

\item{period}{A numeric value that indicates how long one vital is good for, default is one hour.
Note: This will only fill in missing vitals.}

\item{vitals}{A character vector denoting what columns represent the vitals. It is important to 
remember that right now only dopamine and dobutamin are considered in vasopressors used.Platelets should 
be in 10^3/microliter and FiO2 should be listed as a percent, not a decimal.Furthermore bilirubin
should be in micrograms/decilitre. 
needed to calculate SOFA}
}
\description{
Finds the current organ failure score using SOFA.
}
\examples{
\dontrun{
result <- find_sofa(.data = your_data, pateintid = Encounter, time = Service_Timestamp, 
 vitals = c("PaO2" = "Pa02", "FiO2" = "FiO2", 
 "Platelets" = "Platelets", "Bilirubin" = "Bilirubin", "GCS" = "GCS", 
 "Creatinine" = "Creatinine", "SBP" = "SBP", "DBP" = "DBP", "Vasopressor" = "Vasopressor", 
 "Vasopressor_dose" = "Vasopressor Dosage"))
}
}
