% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_baseline_function.R
\name{get_baseline}
\alias{get_baseline}
\alias{get_baseline.bili}
\alias{get_baseline.plate}
\title{Function to get the baseline lab value}
\usage{
get_baseline(.data, labvalue, patientid, result_day, onset_type, blood)

\method{get_baseline}{bili}(.data, labvalue, patientid, result_day, onset_type, blood)

\method{get_baseline}{plate}(.data, labvalue, patientid, result_day, onset_type, blood)
}
\arguments{
\item{.data}{Your dataset}

\item{labvalue}{Your lab value, can be bilirubin, creatinine, GFR, or platelets}

\item{patientid}{A column denoting a unique patient identifier}

\item{result_day}{A Date column denoting when a lab resulted}

\item{onset_type}{A character value denoting if an encounter is community or hospital onset}

\item{blood}{A POSIXct indicating the blood culture time}
}
\value{
Returns a dataset with a new column for the baseline of the specified lab value
}
\description{
Function to get the baseline lab value

baseline function for bilirubin and creatinine

baseline function for GFR and Platelets
}
\examples{
\dontrun{
data <- get_baseline(creat, result_day = lab_results, .blood = blood_data)
}
}
