% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sirs_data}
\alias{sirs_data}
\title{sirs dataset}
\format{A data frame with 9 rows and 6 variables:
\describe{
  \item{Encounter}{A unique patient id}
  \item{Service_Timestamp}{Timestamps of when a variable was taken}
  \item{HR}{Heart Rate, A list of heart rate readings}
  \item{Temperature}{A list of temperature readings in Celsius}
  \item{RR}{Respiratory Rate, A list of respiratory rate readings}
  \item{Wbc}{White Blood Cell, A list of white blood cell readings in mmm^3}
  ...
}}
\usage{
sirs_data
}
\description{
A mock-up dataset to test the sirs organ failure score for one patient.
}
\keyword{datasets}
