% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sofa_data}
\alias{sofa_data}
\title{sofa dataset}
\format{A data frame with 9 rows and 12 variables:
\describe{
  \item{Encounter}{A unique patient id}
  \item{Service_Timestamp}{Timestamps of when a variable was taken}
  \item{Pa02}{Partial Pressure of oxygen, A list of Pa02s}
  \item{Fi02}{Fraction of inspired oxygen, A list of Fi02s}
  \item{Platelet}{A list of platelet counts}
  \item{Bili}{Bilirubin, A list of Bilirubin levels}
  \item{GCS}{Glascow Coma Score, A list of GCSs}
  \item{Creatinine}{A list of creatinine levels}
  \item{SBP}{Systolic Blood Pressure, A list of SBPs}
  \item{DBP}{Dyastolic Blood Pressure, A list of DBPs}
  \item{Vasopressor}{A list of vasopressors given to a patient}
  \item{Vasopressor Dosage}{A list of corresponding dosages of vasopresssors given to a patient}
  ...
}}
\usage{
sofa_data
}
\description{
A mock-up dataset to test the sofa organ failure score for one patient.
}
\keyword{datasets}
