% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_baseline_function.R
\name{get_baseline}
\alias{get_baseline}
\alias{get_baseline.bili}
\alias{get_baseline.plate}
\title{Function to get the baseline lab value}
\usage{
get_baseline(.data, labvalue, patientid, result_day, .blood = NULL)

\method{get_baseline}{bili}(.data, labvalue, patientid, result_day, .blood = NULL)

\method{get_baseline}{plate}(.data, labvalue, patientid, result_day, .blood = NULL)
}
\arguments{
\item{.data}{Your dataste}

\item{labvalue}{Your lab value}

\item{patientid}{A column denoting a unique patient identifier}

\item{result_day}{A Date column denoting when a lab resulted}

\item{.blood}{The blood dataset you will use to calculate window period}
}
\description{
Function to get the baseline lab value

baseline function for bilirubin and creatinine

baseline function for GFR and Platelets
}
\examples{
\dontrun{
data <- get_baseline(creat, .blood = blood_data)
}
}
